sample/AgenticTranscoder/pyproject.toml
[project]
name = "agentic-transcoder"
version = "0.1.0"
description = "Multi-agent code transformation - AgentBuilder to Agentic Flow"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agentic-flow",
    "openai-agents>=0.3.2",
    "rich>=13.0",
    "fire>=0.7",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "black>=25.12.0",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "chatkit",
]

[tool.uv.sources]
agentic-flow = { path = "../..", editable = true }

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.14",
]

[project.scripts]
transcoder = "agentic_transcoder.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_transcoder"]
include = ["src/agentic_transcoder/template/**/*"]

[tool.hatch.build.targets.sdist]
include = ["src/agentic_transcoder", "src/agentic_transcoder/template/**/*"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E402"]  # Allow imports after sys.path manipulation

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src", "../../src"]
testpaths = ["tests"]
markers = [
    "integration: requires real API calls (skipped in CI)",
    "slow: full pipeline tests (manual execution only)",
]

# CI runs only structure + mock tests (fast, deterministic)
# pytest tests/test_structure.py tests/test_mock.py

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-timeout>=2.3.1",
    "ruff>=0.14.10",
]
